Fix issues

+++ src/auth/auth_service.js
@@ -5,6 +5,7 @@ const Errors = require('../common/exceptions');
 const moment = require('moment');
 
 exports.createOtp = async (mobileNumber, roleId) => {
+    console.log(`create otp, mobile = ${mobileNumber}, role id = ${roleId}`);
     if (mobileNumber === undefined || mobileNumber === '' || roleId === undefined || roleId === '')
         throw new Errors.InvalidInputException();
 
@@ -22,23 +23,21 @@ exports.createOtp = async (mobileNumber, roleId) => {
 };
 
 exports.verifyOtp = async (mobileNumber, code) => {
+    console.log(`verify otp, mobile = ${mobileNumber}, otp = ${code}`);
     if (mobileNumber === undefined || mobileNumber === '' || code === undefined || code === '')
-        throw new Errors.InvalidInputException();
+        throw new Errors.InvalidInputException("invalid mobile");
 
     let otp = await authRepository.getByCode(code);
     let user = await userService.findUser(null, mobileNumber);
 
     if (!otp || !user)
-        throw new Errors.InvalidInputException();
+        throw new Errors.InvalidInputException("invalid otp or user");
 
     if (otp.user_id !== user.id)
-        throw new Errors.UnauthorizedException();
-
-    if (otp.used === true)
-        throw new Errors.InvalidInputException();
+        throw new Errors.UnauthorizedException("otp not for user");
 
     if (moment(otp.expires_at).isBefore(moment()))
-        throw new Errors.UnauthorizedException();
+        throw new Errors.UnauthorizedException("otp is expired");
 
     await authRepository.setCodeUsed(otp.id);
 
@@ -69,9 +68,15 @@ exports.authenticateToken = async (token) => {
 };
 
 exports.logout = (token) => {
+    console.log(`logout usecase, token = ${token}`);
+    if (token === '' || token === null || token === undefined)
+        throw new Errors.InvalidInputException();
+
     return authRepository.deleteToken(token);
 };
 
 function generateOtp() {
+    if (process.env.NODE_ENV === 'local')
+        return 1234;
     return Math.floor(1000 + Math.random() * 9000);
 }
